+++ cmd/som/cli/plot_density.go
@@ -3,16 +3,18 @@ package cli
 import (
 	"github.com/mlange-42/som"
 	"github.com/mlange-42/som/plot"
+	"github.com/mlange-42/som/table"
 	"github.com/spf13/cobra"
 	"gonum.org/v1/plot/plotter"
 )
 
-func densityCommand() *cobra.Command {
+func plotDensityCommand() *cobra.Command {
 	var size []int
 	var dataFile string
 	var labelsColumn string
 	var delim string
 	var noData string
+	var ignore []string
 
 	command := &cobra.Command{
 		Use:   "density [flags] <som-file> <out-file>",
@@ -25,10 +27,10 @@ func densityCommand() *cobra.Command {
 
 			return plotHeatmap(size,
 				somFile, outFile, dataFile,
-				labelsColumn, delim, noData, "Density of data",
-				func(s *som.Som, p *som.Predictor) plotter.GridXYZ {
+				labelsColumn, delim, noData, "Density of data", ignore,
+				func(s *som.Som, p *som.Predictor, r table.Reader) (plotter.GridXYZ, []string, error) {
 					density := p.GetDensity()
-					return &plot.IntGrid{Size: *s.Size(), Values: density}
+					return &plot.IntGrid{Size: *s.Size(), Values: density}, nil, nil
 				},
 			)
 		},
@@ -37,6 +39,7 @@ func densityCommand() *cobra.Command {
 	command.Flags().IntSliceVarP(&size, "size", "s", []int{600, 400}, "Size of individual heatmap panels")
 	command.Flags().StringVarP(&dataFile, "data-file", "f", "", "Data file. Required")
 	command.Flags().StringVarP(&labelsColumn, "labels", "l", "", "Labels column in the data file")
+	command.Flags().StringSliceVarP(&ignore, "ignore", "i", []string{}, "Ignore these layers for BMU search")
 
 	command.Flags().StringVarP(&delim, "delimiter", "d", ",", "CSV delimiter")
 	command.Flags().StringVarP(&noData, "no-data", "n", "", "No-data value (default \"\")")