+++ cmd/som/cli/plot_error.go
@@ -3,17 +3,19 @@ package cli
 import (
 	"github.com/mlange-42/som"
 	"github.com/mlange-42/som/plot"
+	"github.com/mlange-42/som/table"
 	"github.com/spf13/cobra"
 	"gonum.org/v1/plot/plotter"
 )
 
-func errorCommand() *cobra.Command {
+func plotErrorCommand() *cobra.Command {
 	var size []int
 	var dataFile string
 	var labelsColumn string
 	var delim string
 	var noData string
 	var rmse bool
+	var ignore []string
 
 	command := &cobra.Command{
 		Use:   "error [flags] <som-file> <out-file>",
@@ -33,10 +35,10 @@ func errorCommand() *cobra.Command {
 
 			return plotHeatmap(size,
 				somFile, outFile, dataFile,
-				labelsColumn, delim, noData, title,
-				func(s *som.Som, p *som.Predictor) plotter.GridXYZ {
+				labelsColumn, delim, noData, title, ignore,
+				func(s *som.Som, p *som.Predictor, r table.Reader) (plotter.GridXYZ, []string, error) {
 					mse := p.GetError(rmse)
-					return &plot.FloatGrid{Size: *s.Size(), Values: mse}
+					return &plot.FloatGrid{Size: *s.Size(), Values: mse}, nil, nil
 				},
 			)
 		},
@@ -47,6 +49,7 @@ func errorCommand() *cobra.Command {
 	command.Flags().IntSliceVarP(&size, "size", "s", []int{600, 400}, "Size of individual heatmap panels")
 	command.Flags().StringVarP(&dataFile, "data-file", "f", "", "Data file. Required")
 	command.Flags().StringVarP(&labelsColumn, "labels", "l", "", "Labels column in the data file")
+	command.Flags().StringSliceVarP(&ignore, "ignore", "i", []string{}, "Ignore these layers for BMU search")
 
 	command.Flags().StringVarP(&delim, "delimiter", "d", ",", "CSV delimiter")
 	command.Flags().StringVarP(&noData, "no-data", "n", "", "No-data value (default \"\")")